get rid of italics, other tweaks

+++ ooApp/OOAPI.m
@@ -711,7 +711,7 @@ NSString *const kKeySearchFilter = @"filter";
         restaurantIDs = [NSMutableArray array];
         [restaurants enumerateObjectsUsingBlock:^(id  _Nonnull obj, NSUInteger idx, BOOL * _Nonnull stop) {
             RestaurantObject *ro = (RestaurantObject *)obj;
-            [restaurantIDs addObject:[NSString stringWithFormat:@"%lu",(unsigned long)ro.restaurantID]];
+            [restaurantIDs addObject:[NSString stringWithFormat:@"%lu", ro.restaurantID]];
         }];
     }
     UserObject *userInfo= [Settings sharedInstance].userObject;
